move test/run -> test/lib/run-spec
