+++ test/handlebars.spec.js
@@ -1,7 +1,7 @@
 'use strict';
 const format = require('../format');
 const path = require('path');
-const runSpec = require('./run');
+const runSpec = require('./lib/run-spec');
 
 describe('handlebars format', function() {
   runSpec(
